+++ TinySQLD-B/QueryProcessor/Operations/SetDatabase.cs
@@ -1,0 +1,16 @@
+﻿using Entities;
+
+namespace QueryProcessor.Operations
+{
+    internal class SetDatabase
+    {
+        public SetDatabase()
+        {
+        }
+
+        internal OperationStatus Execute(string dbName)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
