+++ TinySQLD-B/QueryProcessor/Parser/SQLValidator.cs
